requirements.txt
streamlit==1.38.0
pymupdf==1.24.10
sentence-transformers==3.0.1
faiss-cpu
google-generativeai==0.8.3
langdetect==1.0.9
transformers==4.45.1
torch==2.4.1
python-dotenv==1.0.1
pandas==2.2.2
numpy==2.1.1
googletrans==4.0.0-rc1
streamlit-option-menu==0.3.12
PyPDF2==3.0.1
